refactor(cli): extract onboard actions
